Fixed issues
