--- Python/README.md
@@ -1,25 +1,0 @@
-# Rock, Paper, Scissors (Python)
-
-This is a simple Rock Paper Scissors game in Python
-
-## Installation
-
-Install Python in your editor. To install Python in Visual Studio Code, head to https://code.visualstudio.com/docs/python/python-tutorial
-
-## Compilation
-
-To compile this code for this application, open a terminal. In the terminal navigate to the rock-paper-scissors directory. Then run the command: 
-
-```
-python.py
-```
-
-## Syntax
-
-This is the syntax to run the file and play the Rock, Paper, Scissors game.
-
-```
-python3 Python/python.py
-```
-
-
