+++ python/README.md
@@ -1,0 +1,17 @@
+# Rock, Paper, Scissors (Python)
+
+This is a simple Rock, Paper, Scissors game in Python
+
+## Installation
+
+Install Python in your editor. To install Python, head to https://www.python.org/downloads. 
+
+## Syntax
+
+This is the syntax to run the file and play the Rock, Paper, Scissors game.
+
+```
+python3 python/rps.py
+```
+
+
